requirements.txt
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
python-dotenv==1.0.1
supabase==2.4.5
